+++ MapBrowser/InstalledMap.cs
@@ -2,7 +2,8 @@ using Timberborn.MapItemsUI;
 
 namespace IgorZ.MapBrowser;
 
-sealed record InstalledMap(MapItem Map, string PublishedFileId) {
-  public string Key => Map.MapFileReference.Path ?? Map.MapFileReference.Name;
+sealed record InstalledMap(MapItem Map, string PublishedFileId, WorkshopItemMetadata Metadata = null) {
+  public string Key => Map?.MapFileReference.Path ?? Map?.MapFileReference.Name ?? PublishedFileId;
+  public bool IsInstalled => Map != null;
   public bool Removed { get; set; }
 }